Scope distributed queries to the min-time for each engine
The current algorithm for distributing queries only takes into account
the max-time advertised by each engine. When a remote engine is deployed
in front of Prometheus or Receivers, those stores can apply retention
periods at different times, leading to remote aggregations getting calculated
on a subset of the data. Since samples from newer engines are prioritized
over samples from older engines, these aggregates will end up in the end-result.

To work around the issue, this commit introduces adds a MinTime method to the
remote engine interface. The MinTime is used to scope remote queries at the lower
end so that partial data is never aggregated. The correct value for MinTime should
be determined by the remote engine itself. In the case of Thanos, we can use the
highest min-time from all TSDB backed by a given store.
