+++ logicalplan/propagate_selectors.go
@@ -14,7 +14,7 @@ import (
 // two vector selectors in a binary expression.
 type PropagateMatchersOptimizer struct{}
 
-func (m PropagateMatchersOptimizer) Optimize(expr parser.Expr) parser.Expr {
+func (m PropagateMatchersOptimizer) Optimize(expr parser.Expr, _ *Opts) parser.Expr {
 	traverse(&expr, func(expr *parser.Expr) {
 		binOp, ok := (*expr).(*parser.BinaryExpr)
 		if !ok {
